requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
fastembed==0.4.1
numpy==1.26.4
python-dotenv==1.0.1